+++ src/main/java/com/github/L_Ender/cataclysm/client/particle/Options/TrackLightningParticleOptions.java
@@ -14,9 +14,9 @@ public record TrackLightningParticleOptions(int r, int g, int b) implements Part
     //For networking. Encoder/Decoder functions very intuitive
     public static StreamCodec<? super ByteBuf, TrackLightningParticleOptions> STREAM_CODEC = StreamCodec.of(
             (buf, option) -> {
-                buf.writeFloat(option.r);
-                buf.writeFloat(option.g);
-                buf.writeFloat(option.b);
+                buf.writeInt(option.r);
+                buf.writeInt(option.g);
+                buf.writeInt(option.b);
             },
             (buf) -> {
                 return new TrackLightningParticleOptions(buf.readInt(), buf.readInt(), buf.readInt());
